EmotionAIApp/requirements.txt

streamlit
opencv-python
tensorflow
keras
pillow
numpy
